bug epoll event trigger wrong!!!

+++ myRPC/net/fd_event_group.cc
@@ -32,7 +32,7 @@ FdEventGroup::~FdEventGroup() {
 
 FdEvent* FdEventGroup::getFdEvent(int fd) {
     ScopeMutex<Mutex> groupLock(m_mutex);
-    if(fd < m_fd_group.size()) {
+    if((size_t)fd < m_fd_group.size()) {
         // this fd exist
         return m_fd_group[fd];
     }
